waifu-script/requirements.txt
pyyaml
requests
Pillow
colorthief